Extract recognizers-base submodule
