+++ JavaScript/package.json
@@ -68,6 +68,7 @@
 		"express": "^4.16.1"
 	},
 	"dependencies": {
+		"recognizers-text-base": "file:./packages/recognizers-base",
 		"recognizers-text-number": "file:./packages/recognizers-number",
 		"recognizers-text-number-with-unit": "file:./packages/recognizers-number-with-unit",
 		"recognizers-text-date-time": "file:./packages/recognizers-date-time"
