+++ JavaScript/packages/recognizers-base/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "recognizers-text-base",
+  "version": "1.0.0-rc.27",
+  "description": "recognizers-text-base README",
+  "author": "Microsoft Corp.",
+  "license": "MIT",
+  "keywords": [
+    "nlp",
+    "nlp-entity-extraction",
+    "entity-extraction",
+    "parser-library"
+  ],
+  "main": "dist/recognizers-text-base.umd.js",
+  "module": "dist/recognizers-text-base.es5.js",
+  "typings": "dist/types/recognizers-text-base.d.ts",
+  "browser": "dist/recognizers-text-base.browser.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Microsoft/Recognizers-Text"
+  },
+  "bugs": {
+    "url": "https://github.com/Microsoft/Recognizers-Text/issues"
+  },
+  "engines": {
+    "node": ">=6.0.0"
+  },
+  "scripts": {
+    "clean-build": "rimraf dist && rimraf compiled",
+    "prepare": "npm run clean-build && tsc && rollup -c"
+  }
+}
