+++ JavaScript/packages/recognizers-base/src/recognizer.ts
