+++ JavaScript/packages/recognizers-base/src/recognizers-text-base.ts
@@ -1,0 +1,3 @@
+export { Culture, CultureInfo } from "./culture";
+export { IModel, ModelResult, ModelContainer } from "./models";
+export { IRecognizer, Recognizer } from "./recognizer";
