+++ JavaScript/packages/recognizers-base/tsconfig.json
@@ -1,0 +1,19 @@
+{
+    "compilerOptions": {
+        "module": "commonjs",
+        "target": "es2015",
+        "outDir": "compiled",
+        "sourceMap": true,
+        "rootDir": "src",
+        "moduleResolution": "node",
+        "declaration": true,
+        "declarationDir": "dist/types",
+        "allowSyntheticDefaultImports": true,
+        "typeRoots": [
+            "node_modules/@types"
+        ]
+    },
+    "include": [
+        "src"
+    ]
+}
