+++ JavaScript/packages/recognizers-date-time/package.json
@@ -35,6 +35,7 @@
   "dependencies": {
     "lodash.isequal": "^4.5.0",
     "lodash.tonumber": "^4.0.3",
+    "recognizers-text-base": "^1.0.0-rc.27",
     "recognizers-text-number": "^1.0.0-rc.27",
     "recognizers-text-number-with-unit": "^1.0.0-rc.27"
   }
