+++ JavaScript/packages/recognizers-number-with-unit/package.json
@@ -34,6 +34,7 @@
     "lodash.escaperegexp": "^4.1.2",
     "lodash.last": "^3.0.0",
     "lodash.max": "^4.0.1",
+    "recognizers-text-base": "^1.0.0-rc.27",
     "recognizers-text-number": "^1.0.0-rc.27"
   }
 }
