+++ JavaScript/packages/recognizers-number-with-unit/src/numberWithUnit/numberWithUnitRecognizer.ts
@@ -1,4 +1,5 @@
-import { IModel, Recognizer, Culture, IExtractor, IParser } from "recognizers-text-number";
+import { IModel, ModelResult, Recognizer } from "recognizers-text-base";
+import { Culture, IExtractor, IParser } from "recognizers-text-number";
 import { CurrencyModel, TemperatureModel, DimensionModel, AgeModel } from "./models";
 import { NumberWithUnitExtractor } from "./extractors";
 import { NumberWithUnitParser } from "./parsers";
