+++ JavaScript/packages/recognizers-number/package.json
@@ -35,6 +35,7 @@
     "lodash.escaperegexp": "^4.1.2",
     "lodash.sortby": "^4.7.0",
     "lodash.trimend": "^4.5.1",
-    "xregexp": "^3.2.0"
+    "xregexp": "^3.2.0",
+    "recognizers-text-base": "^1.0.0-rc.27"
   }
 }
