+++ JavaScript/packages/recognizers-number/src/number/numberRecognizer.ts
@@ -1,5 +1,5 @@
-import { IModel } from "../models";
-import { Recognizer } from "../recognizer";
+import { IModel } from "recognizers-text-base";
+import { Recognizer } from "recognizers-text-base";
 import { Culture } from "../culture";
 import { NumberMode, NumberModel, OrdinalModel, PercentModel } from "./models";
 import { AgnosticNumberParserType, AgnosticNumberParserFactory } from "./agnosticNumberParser";
